solve [-Werror=sign-compare]: change int to uint and add casts

+++ crypto_sign/cross/clean/merkle.c
@@ -54,7 +54,7 @@ static
 void setup_tree(uint16_t layer_offsets[LOG2(T)+1], 
                 uint16_t nodes_per_layer[LOG2(T)+1]) {
     uint32_t depth, layer;
-    int r_leaves;
+    uint32_t r_leaves;
     int subtree_found;
 
     /* Initialize array with full node counts */
@@ -195,7 +195,7 @@ void PQCLEAN_CROSS_CLEAN_generate_merkle_tree(unsigned char merkle_tree[NUM_NODE
     parent_layer = LOG2(T)-1;
     for (i=NUM_NODES_MERKLE_TREE-1; i>0; i -= 2) {
         hash(merkle_tree + OFFSET(PARENT(i) + layer_offsets[parent_layer]), merkle_tree + OFFSET(SIBLING(i)), 2*HASH_DIGEST_LENGTH);
-        if (node_ctr >= nodes_per_layer[parent_layer+1] - 2) {
+        if (node_ctr >= (uint32_t) nodes_per_layer[parent_layer+1] - 2) {
             parent_layer--;
             node_ctr = 0;
         } else {
@@ -254,7 +254,7 @@ void PQCLEAN_CROSS_CLEAN_generate_merkle_proof(uint16_t merkle_proof_indices[TRE
             merkle_proof_indices[(*merkle_proof_len)++] = i;
 
         /* Due to the unbalenced structure we got to keep track of the nodes per layer processed */
-        if (node_ctr >= nodes_per_layer[parent_layer+1] - 2) {
+        if (node_ctr >= (uint32_t) nodes_per_layer[parent_layer+1] - 2) {
             parent_layer--;
             node_ctr = 0;
         } else {
@@ -310,7 +310,7 @@ void PQCLEAN_CROSS_CLEAN_rebuild_merkle_tree(unsigned char merkle_tree[NUM_NODES
 
         /* Both siblings are unused, but it must be kept track of the node and layer counter to chose the right offsets */
         if (flag_tree_valid[i] == INVALID_MERKLE_NODE && flag_tree_valid[SIBLING(i)] == INVALID_MERKLE_NODE) {
-            if (node_ctr >= nodes_per_layer[parent_layer+1] - 2) {
+            if (node_ctr >= (uint32_t) nodes_per_layer[parent_layer+1] - 2) {
                 parent_layer--;
                 node_ctr = 0;
             } else {
@@ -340,7 +340,7 @@ void PQCLEAN_CROSS_CLEAN_rebuild_merkle_tree(unsigned char merkle_tree[NUM_NODES
         hash(merkle_tree + OFFSET(PARENT(i) + layer_offsets[parent_layer]), hash_input, 2*HASH_DIGEST_LENGTH);
         flag_tree_valid[PARENT(i) + layer_offsets[parent_layer]] = VALID_MERKLE_NODE;
 
-        if (node_ctr >= nodes_per_layer[parent_layer+1] - 2) {
+        if (node_ctr >= (uint32_t) nodes_per_layer[parent_layer+1] - 2) {
             parent_layer--;
             node_ctr = 0;
         } else {
